Refactored the trashbag
It is now optimized according to the loops benchmark - instead of
using table.getn we keep a separate counter. The trashbag is not
used often (primarily during initialisation) but it does have a
significant difference overall. There is commented blocks of code
that you can uncomment to trace where invalid calls to the class
are made - makes them easier to track down and clean up.