Update FBM Builder Types
This updates FactoryBuilderManager Builder Types
Why is this Important? 

This clears up lots of factory searches, what I mean by this is that Factories originally would searching every tech tier so a T1 Land Factory would be looking at T2 and T3 Build Options which of course makes no sense and is tons of Build Conditions that just did not need to be checked and lots of builders that need to be searched. 
This allows the FBM to breathe very much on the factories generally and it should allow them to be much quicker in their decision making and process generally. 
I tested this and it all works including AIs not using the new FBM Builder Types because I left the original ones.