Change unit:GetCurrentLayer() to unit.Layer
Each time a unit changes layer the new layer is stored inside the
unit. In my own investigations this appeared to be a reliable change
to GetCurrentLayer as it always returned the right values. It prevents
an engine call in some of the most called functions (such as
OnMotionHorzEventChange) - which is a significant improvement in
comparison to retrieving a table value.