LUA - AI - AI Managers
# to -- for Additional Files
Added Logs for BuilderManager for the Data
Added Nils to ReportFailure and DelayEqualBuildPlatton so they do not add to the Data Tables if they are not called
Added Brain to so some functions as Self.Brain does NOT equal to brain anymore, which eliminate many overloads on data where things were store the entire brain.
Added Buildertype to Manager Functions and such to pass the data to the EM and Engineers
Continued Optimizations with eliminating table inserts and such. 
Backwards Comp-ability with AIs not using new Engineer Buildertypes, will refer to Any if other such BTs are invalid.
Fixed an Issue in UCBC and Unit.lua where Cybran Build Drones were being included in some function's searches for units containing 'category' causing Nil Values to return and such.