Add a marker utilities file to manage markers on a map.
Specifically it allows AI developers / map scripters to work
with markers without having to worrying about underlying
inefficiences. As an example, retreiving the mass markers
on a map is a common operation for AIs. If done through
the base game code it would re-allocate a new table of
markers each time an AI condition manager starts checking
the state of the game. That is quite wasteful. This file keeps
track of previous calls, caching the result.

Supports crazyrush-like maps.