Refactor ping-preventing code
This new solution allows you to ping as long as:
 - You are not an observer.
 - Your current focus army is an ally to you.
 - You are not in a replay.

It also ensures that the name of the original author is displayed
under the ping. This can be easily evaded however - in the end
we have no full control over the UI code and someone can easily
write a UI mod to prevent this basic check.