Rename 'DummyUnit' -> 'InsignificantUnit'
Idea is to not introduce a new category, but re-use one that all
dummy units should have anyhow: categories.INSIGNIFICANTUNIT.

In turn, we can tune the base code to exclude those units in
general as we're not interested in them. They should not impact
the game in any form or shape - except to help fix gameplay
issues.