Tweak the interface slightly
Adds information on how to find the game settings to the share
conditions tooltip. Adds the replay id to the map name tooltip
as I feel showing the share condition is more relevant when
watching a replay (it previously replaced the share condition
with the replay ID during a replay).