Improve scoreboard (#3633)
Adds the share conditions to the scoreboard. Adds tooltip for 
the share condition, explaining what it means. Adds information 
to the share condition tooltip on how to find the remaining 
game options. Adds a tooltip for the map name to include the
map description. The replay ID can be found in the map description.

This last one is a bit unintuitive, but the replay ID is 99% of the time
not relevant information. It would take up UI space if we'd just put 
it there regardless. 

Once the map generator is updated the settings used to generate
the map are set as its map description. That way you can view those
settings in game when required.