Update initialisation files (#3637)
Allows loading `kyros.nxt` on the regular branch. Adds 
proper checks on the extension of maps / mods instead to
prevent loading in rar / zip / scd files as mods. Improves
debugging information: as an example when something is
excluded the user is informed through the log. Introduces
the removal of incompatible shaders. Incompatible here 
means shaders of previous versions of FAF.