Fix filtering insignificant units (#3636)
We made the assumption that categories.INSIGNIFICANTUNIT
was not widely used by mods. This was wrong - hence we still
introduced a new category specifically for units that inherit the
DummyUnit class. These units should not be reachable by code
as they are not a typical unit - they lack all the functions you'd
expect.

Fixes (for example) the anti teleporting mechanic of Blackops.