Remove falling tree logic / decal creation
The code was causing problems with various units that relies on
multiple damage instances. Example being bombers that use damage
over time or the Aeon t3 strategic bomber. Also caused issues
with the (sometimes severe) amount of decals. As an example: the
decals underneath mass extractors would no longer be rendered.
And last it caused various issues with mods where a projectile
was re-used but no longer has a radius. The call to DamageArea
would crash with a radius of 0.

These changes are in favor for a more generic approach in the
Projectile class. It prevents code duplication and allows us to
write a general approach that is more efficient, for the same
results.