Remove inconsistency with spawning units through the cheat menu
The previous implementation checked every 0.1 seconds if the
mouse was pressed - if it was, then a callback would trigger. In
practice however there can be the case that a mouse click is
shorter than 0.1 seconds, effectively missing a click.

The new implementation uses the build mode introduced by Balthazar
as a callback mechanism - the build preview is considered issueing
a command and we make use of that. In turn, no clicks are missed
while retaining the current functionality.