Optimize marker usage (#3387)
* Add a marker utilities file to manage markers on a map.

Specifically it allows AI developers / map scripters to work
with markers without having to worrying about underlying
inefficiencies. As an example, retrieving the mass markers
on a map is a common operation for AIs. If done through
the base game code it would re-allocate a new table of
markers each time an AI condition manager starts checking
the state of the game. That is quite wasteful. This file keeps
track of previous calls, caching the result. Supports adaptive
and crazy rush-like maps.

Introduces a UI that allows you to inspect the cached 
markers. The window is toggled with 'K' by default.