Dynamic LOD settings (#3662)
Computes the LOD cut off values of props based on its blueprint properties. A prop that occupies less screen space will have a lower cut off value - allowing it to be culled sooner. This improves the framerate of the game in general, while having a minor impact on visual fidelity. The values set in the blueprint are ignored.