Dynamic AI threat calculations (#3660)
The AI threat values have not been updated over the years - they are now computed based on unit statistics during the blueprint loading phase. Introduces a utilities window that allows you to inspect AI-related blueprint values of a unit. You need to set a hotkey for that window. You can find it in the hotkeys dialogue (F1) by searching for 'information'. Cheats need to be enabled in order for the window to show.