Refactor various utility functions
Refactor the following list of getters / setters:

 SetCanTakeDamage
 CheckCanTakeDamage
 CheckCanBeKilled
 SetCanBeKilled
 SetDead
 IsDead
 GetCachePosition

Add a deprecation warning for all of them.

Move all deprecated functions to the bottom of the unit class. Include other deprecated functions that were not moved there before.

Add a deprecated functions header (a comment that says all functions below it are deprecated).
Replace all known references to the list of getters / setters with their actual value instead.