Introduction of profiler (#3674)
Introduces a profiler that is able to count function calls. Is accompanied by an intuitive interface to navigate the mountains of data that the profiler introduces. Enabling the profiler is costly for the simulation and is therefore limited to a game with AIs, a game with cheats enabled or a game with a game developer. Moderators can act on abuse of the profiler.