Remove sound entities from units, make an exception for submarines
According to commit dfb363dcdc6685bdf4daf8f94178a7da2b06ae8f the
sound of submarine units do not work. The solution introduced is
applied to all units and allocates a handful of entities per
unit. These allocations are expensive!

By default, units try to play the sound themselves. For
submarine units we introduce one entity that plays all
the sounds of the submarine. This appears to work fine.