Optimize shield interactions v3 (#3699)
Optimizes shield interactions by reducing the garbage generated and by removing the 'alive' thread that would run every tick. A novel callback mechanism is introduced that allows a unit to listen when a player has no energy left in storage, and when it has 'a sustainable amount of energy' in storage. This introduces a small balance change: shields no longer attempt to turn on when you have no power. Instead, they start regenerating when you have 10% or more energy in storage.