Generalize knocking trees and creation of decals (#3719)
Improves the performance surrounding the 'tree knocking' mechanic and the creation of splats by projectiles. Prevents excessive creation of splats and terrain effects using a similar mechanic to that of shields introduced by #3681 . Prevents base game effects from having an LOD set to infinity, improving the simulation speed and your framerate.

Closely related to the game development session of the 24th of March.