Refractor CreateVeterancyBuffs
Closes: #3580

Moves the (repeated) allocation of two tables to the global scope. This prevents
the useless repeated reallocation and garbage collection of the data structures.
From the game perspective nothing should change, from a performance POV this should
have limited effect, as this function is only called once per unit type and vet
level.