Improved Trashbag management (#3743)
Trashbags are used to keep track of instances that need to be destroyed at some point. The code mixes the use of Trashbags and raw tables for this task. They now all use the Trashbag for this functionality, preventing a significant amount of table allocations.

Partially reverts #3332 that allowed trashbags to become arbitrarily large in memory.

Tested with the SCTA mod to confirm stability.