InLobbyAutoBalancing
New AutoBalancing function written from scratch - it works in lobby
Added Autobalancing button and its 20 images (because faction color-coding/button highlighting/etc)...
Moved lobby buttons to accommodate the new button

What it does: it balances all occupied slots into two teams with equal numbers of players.  If half of the occupied slots are set to team 1 and half to team 2, then it balances the players while keeping the team-slot matches.  If the teams are not set that way, they are changed automatically to be alternating team 1 and team 2.  If there are an odd number of occupied slots, the last one is set to team - (no team) and the others are balanced without it.

How it balances: this function checks every possible balance combination for making the two teams (while keeping their player counts equal to half the number of occupied slots, rounded down, and not using the last player if there is an odd number of players).  To do this, the function sums up all the relevant players' ratings (keeping mean and deviation separate), and then divides by two.  That yields the goal value for each team.  Any deviation from that is calculated to be the team's imbalance value.  Then, the various team combinations are tested, and the one with the lowest imbalance value is used.

Note that the process used by this function to check the different combinations uses relatively efficient logic that minimizes the number of iterations, while still checking every possible balance combination.  For comparison, a previous PR that attempted in-lobby autobalancing resulted in about 490 million iterations for 12 players.  The number of iterations for 12 players in this PR is 462 (not million), and it doesn't sacrifice balance quality to achieve that!

Also note that the balance values used in this are tentative and may be adjusted in the future.