Further optimize abstract projectile classes (#3847)
Further reduces table allocations and improves access patterns for projectiles. Introduces a clear structure to signify to a developer which parts are deprecated, called by Lua or called by the engine. A minor clean up across the projectile hierarchy to remove empty calls and to simplify some of the logic.