Update the shield bar more often (#3860)
Fixes an issue introduced by #3699 . The shield bar (the blue bar below the shield) was not being updated often enough. In turn, the shield bar was not always in sync with the actual health of the shield. This was most notable when the shield is being assisted, right after it took damage and when it was done healing (when being assisted).

This was purely a visual glitch, the shield was being healed and it was in fact full health.