Mute Main Menu ambient music with /nomovie cmdline
This should be controlled with this switch, since ambient sounds were meant
to provide sound for background movie, which is off.
This switch is not normally used by players, since it removes the radio relations
from HQ or others factions during Campaign as well.
This is rather for dev sessions.