Slightly increase guard scan radius of t1 / t3 interceptors (#3974)
As per the executable deployment the guard scan radius is only applied when it is used: during an attack move or a patrol command. This adds some leeway performance-wise, hence we increase the guard scan radius from 33 to 45 for all t1 / t3 interceptors. The old value was set at a 100, causing significant performance issues.