Fix error when transport dies before it picks up units
If the transport was coming to pick up the units and dies on the way, this logic failed since `EntityCategoryContains(categories.TRANSPORTATION, unit)` didnt work on the transports anymore, apparently dead units don't have the categories anymore?
Anyway, it's easier to check just the land units, since we already have them filtered out.