Improved Mod Manager UI (#4018)
Improves the mod manager by introducing a search bar, being able to mark mods as favorite and being able to sort mods on their type. Introduces the capability of interpreting URLs from the `mod_info.lua` file, allowing you to send people to a github, gitlab, bitbucket, sourceforge page or to a topic on the forums.