Adjust collision hitboxes of all transports (#4078)
Each transport had a collision box that extended far beyond the bottom of the transport. This was likely to catch all fire that the cargo could catch instead. We think this is a simulated game, therefore if some cargo ends up soaking some damage for the transport then that is fine. This is primarily relevant for non-area based damage, such as T1 AA.