Fix the save / load functionality of the game (#4112)
An attempt to tackle the save / load functionality of the game. It appears to work again, but requires more testing to be sure. If your game no longer saves and then loads back in then make a bug report.

Note that old save games will not be compatible, as with any release.