Reduce projectile speed of air to air units (#4113)
Allows interceptors to more reliably fire and hit air units that have landed, specifically landed transports. Otherwise the projectiles end up going so fast that they collide with both the air unit and the ground. Due to the collision bug, the game takes the last collision and therefore the interceptors 'miss' their target.