Alternative CPU benchmark pt. 1 (#4114)
Introduces an alternative to tracking / measuring the performance of the computer of a player. The game checks your local performance throughout the game in an interval basis. It keeps track of your personal sim rate and the current number of units. All players will be gathering this data until the next release. In the next release we'll use this data to replace the CPU score, without loss of control