Using the DifferentialUpgradeCostCalculationCost for Radar
using the `DifferentialUpgradeCostCalculationCost` for Radar to make the reclaim more sense the Mass / Energy Values can be debated by the balance team. currently I've just added the previous tier to the next to have the same cost as they are now if uses the upgrade mechanic and makes more expensive if built straight.