Fix an odd CommandMode related bug with queueing factory orders (#4267)
Fixes the issue where if you issue (heh) the command to produce a unit the command mode would end in an indefinite loop, and by doing so it would cancel the first order you try to give. This fixes that by properly resetting the command mode state, which didn't happen previously.