Introduce generated navigational mesh to interpret maps (#4270)
A navigational mesh is used by the computer to be able to understand its environment. Usually this was done via markers, but as not all maps have those we had to come up with an alternative. This is that alternative, allowing AIs to path towards their targets. Eventually more features will be added such as taking into account threat, pre-caching various attack positions for extractors and generating other interesting pieces of information such as where expansions are.