Adjust factory behavior when building and finished building a unit (#4287)
Allows units to wrap to the correct exit as they are finished being constructed. Reduces the roll off time for the average unit. Instead of waiting for the first order to finish (the 'roll off order'), the factory can start production again the moment the unit left the build location.