Fix edge cases with Colossus Tractor Arms(#4337)
Units that are tractored into the water get killed now. The colossus will also try and actively prefer beaming units that are not tractored, the only exception when this doesn't happen is when the player specifically orders the colossus to attack a given unit.

In the 'usual' cases (where you just walk) it works better now.