Allow lower aix cheat multipliers
Description of changes:
This PR changes the range of selectable aix cheat values to include some negative multipliers. 0.5->0.9.

This also required a change to the default selection value as 5 entries are being added to the start of the list. The same default of 1.5 is retained.

This will allow players to choose a harder cheating aix in terms of resources, but lowering their build speed to allow more time for the player to react. Alternatively a player could restrict an aix's resources but retain its build speed.