Fix bugs surrounding the Seraphim sniper (#5537)
Previously it saved the next firing mode's BP and immediately applied only its fire rate to the weapon, and only switched the rest of the stats immediately after firing or if the switch modes button was pressed when the weapon was fully reloaded. This lead to unexpected behavior from the player's point of view.

Now it switches stats immediately, except for movement speed which is synced with the reload time, since the reload and movement reduction are both drawbacks of sniper mode shots, and without syncing you could ignore the movement penalty by toggling snipe mode off then on in between every shot.