Adjust behavior of the Beetle (#5628)
The beetle will no longer explode on its own when it is given a non-attack command. It will only consider targets when it is being captured, reclaimed, attack move and/or when it is trying to attack a target. Intermediate targets may cause the beetles to explode prematurely. 

Improves the effects of the explosion. The explosion is now visible from afar and it generates debris that is on fire.