Re-implement the Billy
The Billy has an arm period of about 2.5 seconds. If the Billy
collides before it is armed then it does not explode. Instead, a
little bit of debris is spawned in to help the player to understand
what happened.

The damage of the Billy is adjusted. It is now similar to the damage
pattern of the Ahwassa. It deals the full 12.000 damage in the
most inner circle (30% of radius). Between 30% to 60% of the radius it
deals 8250 damage. Between 60% to 90% of the radius it deals 4500 damage
and in between 90% to 100% of the radius it deals 750 damage.

The effects of the Billy are slightly adjusted to make it look a bit
different from a regular strategic missile.