Allow custom AI descriptions and ratings
Adds in logic to assign campaign AI personalities to AIs, and to change AI ratings from 0.
For now setup to work with M28AI, with comments/details for if other AIs wished to be added.
Has to be done via the lobby.lua rather than via the AI mod itself, as the AI mod does not load at this point, and hence it is not possible for the AI mod to change these values (and for the changes to then be reflected properly).