Update the use of `STRATEGIC` category (#5942)
The meaning is any high-value unit or structure that deals damage at a large range, that should be protected by the user or sniped by the enemy. Sort of "Anti-Economy" structures/units because they can sit in your base to destroy the enemy economy. Therefore it applies to T2 artillery, Static TMLs, T3 Static Artillery, Nukes, and game enders.