Add tests for rate of fire (#5845)
Not all rate of fire values are actually used by the game. The game is limited to 10 ticks per second. A unit is therefore unable to fire 8 rounds per second - it can either fire 10 rounds per second or 5 rounds per second. This new test suite guarantees that the values that you see in the blueprint represent the values that the game would actually end up using.