Move vetDamage total inside instigator check (#6110)
Previously, all damage taken from any source (including an allied unit
or `self`) would be included in the veterancy dispersal calculations.
This moves the addition operation to `VetDamageTaken` (the denominator
in `veterancyDispersal`) inside an "IsEnemy" check, which prevents (for
example) the Paragon from stealing 10,000 damage of veterancy experience
for itself as it dies.

Downside is a particular unit may get more credit than it is
"rightfully" owed for a unit that was hit by friendly fire--not a big
downside, in the scheme of things.