Enhance filter engineer hotkey (#5249)
The hotkey now considers SACUs and tech 3 engineers to be equivalent.

In addition it will now also filter over factions. As an example, if you have a selection of 5 Aeon engineers and 3 UEF engineers of the same tech then you'll end up selecting the Aeon engineers and the UEF engineers assist an Aeon engineer.