Fix production queue tooltip in non-default layouts (#6243)
## Description of the proposed changes

![image](https://github.com/FAForever/fa/assets/5189292/c4e0bd39-c845-4e08-834d-8625e7eaf315)

The top part of the tooltip is the queue. It can be enabled in
Options->Interface->Additional Information->Show Factory Queue on Hover.

The issue is that the _left and _right layouts are missing the layouting
logic for the queue.

Fix:
* Moved code that handles internal layout/graphics of unit tooltip queue
from unitview_mini.lua SetLayout() to unitview.lua CreateQueueGrid().
This leaves only the two LayoutHelper calls that "dock" the queue
display to the unit tooltip itself. Even though this docking should be
the same for all layouts it makes sense to duplicate them in all
SetLayout() implementations from the readability standpoint.
* Moved the code that assigns themed textures to
controls.queue.SetThemeTextures(). Added a call to SetThemeTextures()
* Duplicated the logic that docks unit tooltip queue and sets theme
textures to the _left and _right layouts.

<!-- A clear and concise description (or visuals) of what the changes
imply. -->
<!-- If it closes an issue, make sure to link the issue by using
"(Closes/Fixes/Resolves) #(Issue Number)" in your pull request. -->


## Testing done on the proposed changes
<!-- List all relevant testing that you've done to confirm the changes
work. -->

Verified the queue tooltip behaves correctly for an observer even when
the layout is cycled
For each layout
 * Selected the layout
 * Restarted the game
 * Started a replay
 * Verified no lazyvar errors appear in logs

Verified the queue tooltip behaves correctly for a player
 * Set the queue tooltip in the options to "Always"
 * Cycled through layouts and everything looked good
* Cycled through themes and found no issues except that on the _right
layout if the theme is cycled after selecting a unit (regardless of
whether it remains selected) in the current theme, there are lazyvar
errors (see the snippet below) and it partially fails (some parts of the
ui update the theme, some don't). This issue is present without the
changes and it looks much worse in the log without the changes. If the
theme is cycled without ever selecting a unit in the current theme, it
works correctly and no lazyvar errors show up.

```
INFO: attempting to set skin to: 	aeon
WARNING: Error running console command UI_RotateSkin +: e:\workshop\coding\faf\lua\lazyvar.lua(207): attempt to call method `OnDirty' (a nil value)
         stack traceback:
         	e:\workshop\coding\faf\lua\lazyvar.lua(207): in function `SetValue'
         	e:\workshop\coding\faf\lua\lazyvar.lua(225): in function `Set'
         	e:\workshop\coding\faf\lua\ui\uiutil.lua(287): in function `SetCurrentSkin'
         	e:\workshop\coding\faf\lua\ui\uiutil.lua(385): in function <e:\workshop\coding\faf\lua\ui\uiutil.lua:350>
```

## Additional context
<!-- Add any other context about the pull request here. -->


## Checklist

- [x] Changes are annotated, including comments where useful
- [x] Changes are documented in the changelog for the next game version