Format the majority of the remaining blueprint files (#6272)
The only blueprint files that are not formatted are the emitter files.
It would merge the key-based and value-based fields, creating a bit of a
mess. Luckily we won't be doing much (if any) batch processing on those
type of files.

With this we can now finally start batch processing file empowered by
the [Brewlan WikiGen](https://github.com/The-Balthazar/BrewWikiGen) made
by @The-Balthazar. This would empower us to do various edits to units,
props and projectiles an mass and bake them right into the blueprint
files.