AI - Fix incorrect unit set in T2LandAmphibious factory build template (#6286)
## Description of the proposed changes
This PR fixes a small error in the AI platoon build templates. The
T2LandAmphibious was set to build the Aeon Aurora instead of the Blaze.


## Testing done on the proposed changes
Validated that the correct unit was being built post change


## Additional context
According to the file history this error has been there for YEARS :D I
even went back and checked the original SCFA lua files, the error was
present then too.


## Checklist
- [x] Changes are annotated, including comments where useful
- [ ] Changes are documented in the changelog for the next game version