Rotate every other texture layer (#5810)
I discovered that rotating some texture layers makes the repeating
texture pattern less visible.
Now not all textures are aligned in the same direction and the 30°
rotation from the horizonal/vertical grid also seems to make it harder
for the eye to notice repeating features.
The effect is not amazing, but it's almost free, so it still seems like
a good deal.
The only question is, on which shaders do we want to apply that?

Before:
![Screenshot 2024-01-07
015257](https://github.com/FAForever/fa/assets/52536103/358478f2-ce0b-4d33-b2bf-a7a8fdb111fa)

After:
![Screenshot 2024-01-07
015310](https://github.com/FAForever/fa/assets/52536103/db69370e-aa1e-4571-8c1f-2ba2112fef7e)