Re-implement Chrono Dampener to properly use firing cycle + VFX fixes (#5883)
Fixes #2410.
To replicate the issue, start chrono with targeting of a land/sea
surface unit, then walk underwater while staying within its 35 range. It
will stay locked on and never check its own inability to fire
underwater, so it'll keep firing underwater.
This is caused by its custom firing cycle that never goes to the OnFire
event except in the initial target acquisition. The cycle:
`idle -> FireReady -(FireReady.OnFire)-> FiringState (repeat forever)
-(FiringState.OnLostTarget)-> Idle`
### Changes
- RateOfFire actually determines the fire rate now.
- It can no longer fire underwater.
- Categories to stun are generated from the BP's stun buff definition
instead of being coded in the weapon.
- Chrono uses its own range instead of relying completely on the primary
weapon. Now the range upgrades adjust the range of Chrono, like they do
with overcharge.
- The initial stun effects are fixed (target.InitialStunFxApplied was
never reset) and the effect was changed to be a more noticeable flash,
as if the enemy was suddenly hit by a burst of energy. The flash effect
also reduces with distance since the stun is weaker far away.
- If you zoomed out or looked away while Chrono was firing, the large
ring particle wouldn't emit and you'd never see the effect if you looked
back at it. This is changed to always emit because Chrono is rather
important.
- `AIR*MOBILE` is not a valid target restriction, so Chrono was
uselessly firing at aircraft. The target priorities are fixed to only
fire at t1-t3 mobile surface units.