Use LOD instead of always emitting for Chrono visuals (#6295)
<!-- General useful tooling:
- [ScreenToGif](https://www.screentogif.com/): Free, open source screen
recorder that can export to MP4. If the changes are visual, these can
help you tell us exactly what the changes imply!
-->
<!-- Feel free to remove unused parts of this template. -->

## Issue
Due to #5883, Chrono's effects are visible through fog of war.

## Description of the proposed changes
<!-- A clear and concise description (or visuals) of what the changes
imply. -->
<!-- If it closes an issue, make sure to link the issue by using
"(Closes/Fixes/Resolves) #(Issue Number)" in your pull request. -->
Use LOD instead of making the effects always visible. The 1100 LOD makes
it always visible from a max zoomed out view on a 10 km map, and for 20
km maps, visible if about 1/4th of the map is visible. This is
appropriate for the huge AoE Chrono has: it's larger than a nuke's inner
radius. It's still weaker than a nuke, so it doesn't have unlimited LOD.

I used the emitter editor to make the changes so the file format is
changed.

## Testing done on the proposed changes
<!-- List all relevant testing that you've done to confirm the changes
work. -->
Build a chrono ACU and have it groundfire. Zoom out to see the effects
are visible at a reasonable zoom level. Switch to civilian army and turn
on fog of war to check that the effects don't emit if the ACU is not
visible.

## Checklist
- [x] Changes are documented in the changelog for the next game version