Move hidden crash damage out of Ahwassa and CZAR scripts (#6311)
## Issue
Ahwassa/CZAR deals 1000 damage in 5 radius due to a line in its script,
which makes that information unknown to the user because it is not in
the blueprint.

## Description of the proposed changes
Removes the scripted damage and adds it to the blueprint. This results
in:
- high damage area radius: 5 -> 10 (Ahwassa) / 15 (CZAR)
- high damage area damage through Seraphim T3 shields: 1000 (constant) +
2800 (reduced by shields) -> 3800 (reduced by shields)
- The shield choice is relevant because shields can reduce crash damage
by up to 20% of their max HP. That means the max crash damage reduction
in normal games is `21k * 0.2 = 4200`. 3800 damage under shields keeps
the ability to one shot SMDs with crash damage.
- high damage against unshielded targets: remains the same

## Testing done on the proposed changes
<!-- List all relevant testing that you've done to confirm the changes
work. -->
Crashing these air T4s through shields deals expected damage.

## Additional context
Previous balance PR reducing crash damage:
https://github.com/FAForever/fa/pull/4543

## Checklist
- [ ] Changes are documented in the changelog for the next game version