Fix unit view clipping in replays for the mini layout (#6291)
<!-- General useful tooling:
- [ScreenToGif](https://www.screentogif.com/): Free, open source screen
recorder that can export to MP4. If the changes are visual, these can
help you tell us exactly what the changes imply!
-->
<!-- Feel free to remove unused parts of this template. -->

## Description of the proposed changes
Before:

![image](https://github.com/FAForever/fa/assets/5189292/c2f560e5-81af-45ba-bec0-79198e27fe5c)
After:

![image](https://github.com/FAForever/fa/assets/5189292/1fb91af2-d7bf-4581-acad-a731f321c6f1)

I have changed the unit view positioning in replays for the mini layout
to the same as the right layout, which is effectively an identical
layout because they differ only in:
1) The orders control position but there is no orders control in
replays.
2) The clipping issue itself, which I assume is not intended :)

In order to achieve that, I have added a condition that checks whether
the orders control exists.

Additionally, I have moved imports in layouts to the top of the file
since that is tradition, it avoids import duplication in some cases and
it is theoretically faster (it saves up to 30% processing time of the
PositionWindow() method, but that still amounts to nothing in the grand
scheme of things).

No changes to non-replay positioning of the mini layout:

![image](https://github.com/FAForever/fa/assets/5189292/39798d91-47b9-419c-ae93-5d363b6c8b42)

## Testing done on the proposed changes
Replay testing:
 * Ran a replay
 * Switched to a specific player
 * For each layout
    * Verified unit view renders correctly without selection
    * Verified unit view renders correctly while a factory is selected
 * Switched to the observer full view
 * For each layout
     * Verified unit view renders correctly while a factory is selected

Skirmish testing:
 * Started a skirmish as a player
 * For each layout
    * Verified unit view renders correctly without selection
    * Verified unit view renders correctly while a factory is selected 

Skirmish observer testing:
* Same as replay testing except the observer has the orders control so
it looks the same as actually playing

## Additional context
I wondered if Gyle runs into this in his casts but he mostly selects
combat units that do not have the extra bits at the top of the
construction control.


## Checklist
- [x] Changes are annotated, including comments where useful
- [x] Changes are documented in the changelog for the next game version