Show player divisions in scoreboard (#6160)
## Description of the proposed changes
Show the division of players in the scoreboard instead of rating if
available (i.e. in matchmaker games)
Closes #4258 

## Testing done on the proposed changes
Tested them on a matchmaker replay, as well as a normal game to make
sure that the rating still shows up when no divisions are available.

The missing space between division and subdivision is because the added
space in the game options is not present in current replays. This is
just an artefact of the testing process and will be fixed once the
release happened.


![grafik](https://github.com/FAForever/fa/assets/52536103/023f29d1-29fd-4cb9-952d-d136b9f5d86b)


## Checklist

- [x] Changes are annotated, including comments where useful
- [x] Changes are documented in the changelog for the next game version